+++ actionview/CHANGELOG.md
@@ -1,3 +1,14 @@
+*   Extend `ActionView::RecordIdentifier#dom_id` to accept variable arguments and
+    keyword arguments to build an identifier list.
+
+    For example, reference multiple elements from within an aria-labelledby
+    attribute:
+
+        tag.button "Edit", aria: { labelledby: dom_id(@post, [:edit, {custom: true}, :title] }
+        # => <button aria-labelledby="edit_post_1 custom_post_1 title_post_1">Edit</button>
+
+    *Sean Doyle*
+
 ## Rails 6.1.0.rc1 (November 02, 2020) ##
 
 *   Yield translated strings to calls of `ActionView::FormBuilder#button`